Initial commit
- Initial implementation of win_io object. The println and scanln
functions have good test coverage but cursorUp is lacking.

- Initial implementation of python test framework. Framework defines a
test harness base type, some assertion (expect) methods and error types,
and a runner to run any tests (types derived from Test) in a specified
module.

- Initial makefile. Builds the win_io object, and a win_io .so to be
used by a python proxy for testing. Includes a build target to produce
the objects, a test target to build the test proxys and run tests, and a
clean target.

- Initial IO tests. IO tests use the standard "derived from Test" types
to start the tests, but because the affect the console they need to use
shell redirection to test their outcome. Additional scripts were added
that exercise the functions so the test can invoke the command in a
subprocess and read its pipes. Only the println and scanln functions are
covered so far.